+++ modules/oscore/src/osocre_hkdf_info.c
@@ -72,7 +72,7 @@ inline enum err oscore_create_hkdf_info(struct byte_array *id,
 		info_struct._oscore_info_id_context_bstr.len = id_context->len;
 	}
 	info_struct._oscore_info_alg_aead_choice = _oscore_info_alg_aead_int;
-	info_struct._oscore_info_alg_aead_int = aead_alg;
+	info_struct._oscore_info_alg_aead_int = (int32_t)aead_alg;
 
 	info_struct._oscore_info_type.value = (uint8_t *)type_enc;
 	info_struct._oscore_info_type.len = (uint32_t)strlen(type_enc);
